requirements.txt
django
djangorestframework
pandas
numpy
scikit-learn
